handle when a client might return an error unrelated to a request:
Handle the edge case when a client returns an error that is not
related to a request. This was reported from a user having received
a response for a request `id` followed by an error with the same
response `id` but we were no longer looking for that `id` since the
response was already processed. This is on the client, but we
should handle it gracefully within the library without hanging
indefinitely.

- Refactor rpc response validation to ``_utils/validation.py`` and use
  this in `PersistentConnectionProvider` to handle any errors that are
  found in the cache. Refactor applying error and null formatters to
  ``_utils/formatters.py``

- Add a test case that covers a similar scenario to what the user
  reported.

- Add a newsfragment for this change.
